Replace in-memory DB with SQL server Db

+++ GovernmentGrantAPI/GovernmentGrantAPI/Model/FamilyMember.cs
@@ -20,7 +20,7 @@ namespace GovernmentGrantAPI.Model
 
         public OccupationType OccupationType { get; set; }
 
-        public decimal AnnualIncome { get; set; }
+        public double AnnualIncome { get; set; }
 
         public DateTime DOB { get; set; }
     }
